Move LLM API types to server folder
