+++ src/containers/ImageUploader/components/ImageContent.tsx
@@ -18,7 +18,7 @@ import { FormField } from "../../../components/FormField";
 import { FormContent } from "../../../components/FormikForm";
 import { AI_ACCESS_SCOPE } from "../../../constants";
 import { useSession } from "../../../containers/Session/SessionProvider";
-import { AltTextVariables } from "../../../modules/llm/llmApiTypes";
+import { AltTextVariables } from "../../../server/llmApiTypes";
 import { TitleField } from "../../FormikForm";
 import { ImageFormikType } from "../imageTransformers";
